infrastructure/upload_to_gcp/requirements.txt
google-cloud-storage==1.43.0